+++ src/modules/esm.mjs
@@ -1,0 +1,46 @@
+import {dirname, join, sep} from "node:path";
+import {fileURLToPath} from "node:url";
+import {release, version} from "node:os";
+import {createServer as createServerHttp} from "node:http";
+import * as cjs from './files/c.js'
+import * as jsonA from "./files/a.json" with {type: "json"};
+import * as jsonB from "./files/b.json" with {type: "json"};
+
+cjs;
+
+const random = Math.random();
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const jsonAPath = await join(__dirname, "files", "a.json");
+const jsonBPath = await join(__dirname, "files", "b.json");
+let unknownObject;
+
+if (random > 0.5) {
+  unknownObject = jsonA;
+} else {
+  unknownObject = jsonB;
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end("Request accepted");
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log("To terminate it, use Ctrl+C combination");
+});
+
+export { unknownObject, myServer };
